requirements.txt
albumentations==1.4.6
matplotlib==3.8.4
mpmath==1.3.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
open3d==0.18.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
pandas==2.2.1
pillow==10.3.0
plotly==5.21.0
py-cpuinfo==9.0.0
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pypylon==3.0.1
pyquaternion==0.9.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
requests==2.31.0
scikit-image==0.23.2
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
torch==2.2.2
torchvision==0.17.2
triton==2.2.0
ultralytics==8.2.15
urllib3==2.2.1
